Fixes QoS 2 persistence provider, Blob-based state persistence provider, SequenceNumber -> Packet Id conversion
Motivation:
Fix issues with Azure Storage-based persistence providers and sequence number conversion logic

Modifications:
- Azure Table QoS 2 persistence provider uses proper query API to retrieve only relevant row
- QoS 2 persistence provider API is modified to work with SequenceNumber instead of MessageId
-Azure Blob session state persistence provider is modified to update ETag value on in-memory session state object to allow saving it later on without any issue.
- 14 least significant bits are used when converting SequenceNumber into PacketId
- SequenceNumber type is changed to ulong.

Result:
- QoS 2 works
- multiple SUBSCRIBE / UNSUBSCRIBE requests work within a single connection lifetime
- generated Packet Id  is conformant with MQTT spec

+++ src/ProtocolGateway.Core/Messaging/IMessage.cs
@@ -21,9 +21,6 @@ namespace Microsoft.Azure.Devices.ProtocolGateway.Messaging
 
         uint DeliveryCount { get; }
 
-        /// <remarks>
-        /// Only 15 least significant bits will be honored
-        /// </remarks>
-        ushort SequenceNumber { get; }
+        ulong SequenceNumber { get; }
     }
 }
